Add a base data model
